TST: isolate solidmotor unit tests.
